New: Category api added

+++ src/services/recipes.service.ts
@@ -6,16 +6,17 @@ import { isEmpty } from '@utils/util';
 class RecipeService {
   public recipe = new PrismaClient().recipe;
   public ingredient = new PrismaClient().ingredient;
+  public categories = new PrismaClient().category;
 
   public async findAllRecipes(): Promise<Recipe[]> {
-    const allRecipes: Recipe[] = await this.recipe.findMany({ include: { ingredients: {include: {ingredient: true}} } });
+    const allRecipes: Recipe[] = await this.recipe.findMany({ include: { ingredients: { include: { ingredient: true } }, categories: { include: { category: true } } } });
     return allRecipes;
   }
 
   public async findRecipeBySlug(recipeSlug: string): Promise<Recipe> {
     if (isEmpty(recipeSlug)) throw new HttpException(400, "RecipeSlug is empty");
 
-    const findRecipe: Recipe[] = await this.recipe.findMany({ where: { slug: recipeSlug }, include: { ingredients: {include: {ingredient: true}} }});
+    const findRecipe: Recipe[] = await this.recipe.findMany({ where: { slug: recipeSlug }, include: { ingredients: { include: { ingredient: true } } } });
     if (!findRecipe) throw new HttpException(409, "Recipe doesn't exist");
 
     return findRecipe[0];
@@ -26,24 +27,40 @@ class RecipeService {
 
     if (isEmpty(recipeData)) throw new HttpException(400, "RecipeData is empty");
 
-    const { ingredients } = recipeData;
+    const { ingredients, categories } = recipeData;
 
     const checkIfIngredientsIdExist = await this.ingredient.findMany({ where: { id: { in: ingredients.map(ingredient => ingredient.id) } } });
     if (checkIfIngredientsIdExist.length !== ingredients.length) throw new HttpException(409, "One or more ingredients doesn't exist");
 
     const findRecipe: Recipe[] = await this.recipe.findMany({ where: { slug: recipeData.slug } });
+
     if (findRecipe.length > 0) throw new HttpException(409, `This slug ${recipeData.slug} already exists`);
 
-    const createRecipeData: Recipe = await this.recipe.create({ data: { ...recipeData, authorId: authUser.id, ingredients: {
-      create: ingredients.map(ingredient => ({ 
-        amount: ingredient.amount, 
-        ingredient: {
-          connect: {
-            id: ingredient.id
-          }
-        } 
-      }))
-    }}});
+    const createRecipeData: Recipe = await this.recipe.create({
+      data: {
+        ...recipeData,
+        authorId: authUser.id,
+        ingredients: {
+          create: ingredients.map(ingredient => ({
+            amount: ingredient.amount,
+            ingredient: {
+              connect: {
+                id: ingredient.id
+              }
+            }
+          }))
+        },
+        categories: {
+          create: categories.map(categoryId => ({
+            category: {
+              connect: {
+                id: categoryId
+              }
+            }
+          }))
+        }
+      }
+    });
     return createRecipeData;
   }
 
@@ -52,26 +69,47 @@ class RecipeService {
 
     const findRecipe: Recipe = await this.recipe.findUnique({ where: { id: recipeId }, include: { ingredients: true } });
     if (!findRecipe) throw new HttpException(409, "Recipe doesn't exist");
-    
-    const { ingredients } = recipeData;
+
+    const { ingredients, categories } = recipeData;
 
     const checkIfIngredientsIdExist = await this.ingredient.findMany({ where: { id: { in: ingredients.map(ingredient => ingredient.id) } } });
     if (checkIfIngredientsIdExist.length !== ingredients.length) throw new HttpException(409, "One or more ingredients doesn't exist");
-    
-    const updateRecipeData = await this.recipe.update({ where: { id: recipeId }, data: { ...recipeData, ingredients: {
-     deleteMany: {
-      recipeId: recipeId,
-      NOT: ingredients.map(({ recipeId }) => ({ recipeId })),
-     },
-      create: ingredients.map(ingredient => ({
-        amount: ingredient.amount, 
-        ingredient: {
-          connect: {
-            id: ingredient.id
-          }
-        } 
-      }))
-    }}});
+
+    const checkIfCategoriesIdExist = await this.categories.findMany({ where: { id: { in: categories.map(categoryId => categoryId) } } });
+    if (checkIfCategoriesIdExist.length !== categories.length) throw new HttpException(409, "One or more categories doesn't exist");
+
+    const updateRecipeData = await this.recipe.update({
+      where: { id: recipeId }, data: {
+        ...recipeData,
+        ingredients: {
+          deleteMany: {
+            recipeId: recipeId,
+            NOT: ingredients.map(({ recipeId }) => ({ recipeId })),
+          },
+          create: ingredients.map(ingredient => ({
+            amount: ingredient.amount,
+            ingredient: {
+              connect: {
+                id: ingredient.id
+              }
+            }
+          }))
+        },
+        categories: {
+          deleteMany: {
+            recipeId: recipeId,
+            NOT: categories.map(({ recipeId }) => ({ recipeId })),
+          },
+          create: categories.map(categoryId => ({
+            category: {
+              connect: {
+                id: categoryId
+              }
+            }
+          }))
+        }
+      }
+    });
     return updateRecipeData;
   }
 
@@ -82,7 +120,7 @@ class RecipeService {
     const findRecipe: Recipe = await this.recipe.findUnique({ where: { id: recipeId } });
     if (!findRecipe) throw new HttpException(409, "Recipe doesn't exist");
 
-    const deleteRecipeData = await this.recipe.delete({ where: { id: recipeId} });
+    const deleteRecipeData = await this.recipe.delete({ where: { id: recipeId } });
     return deleteRecipeData;
   }
